Refactor screenshot automation behind debug guards
