+++ ShortcutCycle/ShortcutCycle/Services/LaunchAtLoginManager.swift
@@ -9,23 +9,34 @@ import ShortcutCycleCore
 @MainActor
 class LaunchAtLoginManager: ObservableObject {
     static let shared = LaunchAtLoginManager()
+    #if DEBUG
     private static var screenshotOverrideEnabled: Bool?
+    #endif
     
     @Published var isEnabled: Bool {
         didSet {
             if isEnabled != oldValue {
+                #if DEBUG
                 if Self.screenshotOverrideEnabled == nil {
                     updateLoginItem()
                 }
+                #else
+                updateLoginItem()
+                #endif
             }
         }
     }
     
     private init() {
         // Read current status from SMAppService
+        #if DEBUG
         self.isEnabled = Self.screenshotOverrideEnabled ?? (SMAppService.mainApp.status == .enabled)
+        #else
+        self.isEnabled = SMAppService.mainApp.status == .enabled
+        #endif
     }
 
+    #if DEBUG
     static func setScreenshotOverride(isEnabled: Bool?) {
         screenshotOverrideEnabled = isEnabled
         if let isEnabled {
@@ -34,6 +45,7 @@ class LaunchAtLoginManager: ObservableObject {
             shared.refreshStatus()
         }
     }
+    #endif
     
     private func updateLoginItem() {
         do {
@@ -53,10 +65,12 @@ class LaunchAtLoginManager: ObservableObject {
     
     /// Refresh the current status from SMAppService
     func refreshStatus() {
+        #if DEBUG
         if let override = Self.screenshotOverrideEnabled {
             self.isEnabled = override
             return
         }
+        #endif
 
         self.isEnabled = SMAppService.mainApp.status == .enabled
     }
