+++ ShortcutCycle/ShortcutCycle/ShortcutCycleApp.swift
@@ -1,9 +1,4 @@
 import SwiftUI
-import KeyboardShortcuts
-import UniformTypeIdentifiers
-#if canImport(ScreenCaptureKit)
-import ScreenCaptureKit
-#endif
 #if canImport(ShortcutCycleCore)
 import ShortcutCycleCore
 #endif
@@ -781,1362 +776,21 @@ enum ShortcutCycleURLRouter {
 }
 
 
-// MARK: - Screenshot Automation
-
-private enum ScreenshotScene: String {
-    case general
-    case group
-    case backups
-    case hudHorizontal = "hud-horizontal"
-    case hudGrid = "hud-grid"
-    case menuPopover = "menu-popover"
-
-    var defaultGroupKey: ScreenshotFixtureGroupKey? {
-        switch self {
-        case .general, .backups, .menuPopover:
-            return nil
-        case .group:
-            return .utilities
-        case .hudHorizontal:
-            return .info
-        case .hudGrid:
-            return .manyApps
-        }
-    }
-
-    var baseCaptureDelay: TimeInterval {
-        switch self {
-        case .backups:
-            return 1.2
-        case .general, .group:
-            return 0.8
-        case .hudHorizontal, .hudGrid, .menuPopover:
-            return 0.45
-        }
-    }
-
-    var hidesWindowChrome: Bool {
-        switch self {
-        case .hudHorizontal, .hudGrid, .menuPopover:
-            return true
-        case .general, .group, .backups:
-            return false
-        }
-    }
-
-    var showsTrafficLightsOverlay: Bool {
-        switch self {
-        case .general, .group, .backups:
-            return true
-        case .hudHorizontal, .hudGrid, .menuPopover:
-            return false
-        }
-    }
-}
-
-private enum ScreenshotFixtureGroupKey: String {
-    case info
-    case communication
-    case productivity
-    case utilities
-    case media
-    case manyApps = "many-apps"
-}
-
-private enum ScreenshotMenuVariant: String {
-    case `default`
-    case selected
-}
-
-private struct ScreenshotArguments {
-    let scene: ScreenshotScene
-    let theme: AppTheme
-    let language: String
-    let outputURL: URL
-    let windowInfoURL: URL?
-    let backgroundURL: URL?
-    let groupKey: ScreenshotFixtureGroupKey?
-    let menuVariant: ScreenshotMenuVariant
-
-    var prefersLiveWindowCapture: Bool {
-        windowInfoURL != nil
-    }
-
-    var captureDelay: TimeInterval {
-        let themeSettleDelay: TimeInterval
-        switch scene {
-        case .general:
-            themeSettleDelay = theme == .dark ? 0.55 : 0.2
-        case .group:
-            themeSettleDelay = theme == .dark ? 1.05 : 0.35
-        case .backups:
-            themeSettleDelay = theme == .dark ? 0.75 : 0.3
-        case .hudHorizontal, .hudGrid, .menuPopover:
-            themeSettleDelay = 0
-        }
-
-        return scene.baseCaptureDelay + themeSettleDelay
-    }
-
-    static let current = parse(CommandLine.arguments)
-
-    private static func parse(_ arguments: [String]) -> ScreenshotArguments? {
-        func value(for flag: String) -> String? {
-            guard let index = arguments.firstIndex(of: flag),
-                  arguments.indices.contains(index + 1) else {
-                return nil
-            }
-            return arguments[index + 1]
-        }
-
-        guard let rawScene = value(for: "--screenshot-scene"),
-              let scene = ScreenshotScene(rawValue: rawScene),
-              let rawOutput = value(for: "--screenshot-output") else {
-            return nil
-        }
-
-        let theme = AppTheme(rawValue: value(for: "--screenshot-theme") ?? "") ?? .light
-        let language = value(for: "--screenshot-language") ?? "en"
-        let windowInfoURL = value(for: "--screenshot-window-info").map { URL(fileURLWithPath: $0) }
-        let backgroundURL = value(for: "--screenshot-background").map { URL(fileURLWithPath: $0) }
-        let groupKey = value(for: "--screenshot-group").flatMap(ScreenshotFixtureGroupKey.init(rawValue:))
-            ?? scene.defaultGroupKey
-        let menuVariant = value(for: "--screenshot-variant").flatMap(ScreenshotMenuVariant.init(rawValue:))
-            ?? .default
-
-        return ScreenshotArguments(
-            scene: scene,
-            theme: theme,
-            language: language,
-            outputURL: URL(fileURLWithPath: rawOutput),
-            windowInfoURL: windowInfoURL,
-            backgroundURL: backgroundURL,
-            groupKey: groupKey,
-            menuVariant: menuVariant
-        )
-    }
-}
-
-private struct ScreenshotWindowInfo: Encodable {
-    let windowNumber: Int
-}
-
-private enum ScreenshotRuntime {
-    static let windowSize = CGSize(width: 1440, height: 900)
-    static let outputPixelSize = CGSize(width: 2880, height: 1800)
-
-    private static let screenshotShortcutKeys: [KeyboardShortcuts.Key] = [
-        .one, .two, .three, .four, .five, .six
-    ]
-
-    static func primeDefaults(for arguments: ScreenshotArguments) {
-        let defaults = UserDefaults.standard
-        defaults.set(true, forKey: WelcomeExperiencePolicy.hasDismissedWelcomeKey)
-        defaults.set(true, forKey: WelcomeExperiencePolicy.hasAutoOpenedWelcomeSettingsKey)
-        defaults.set(arguments.language, forKey: "selectedLanguage")
-        defaults.set(arguments.theme.rawValue, forKey: "appTheme")
-        defaults.set(true, forKey: "showHUD")
-        defaults.set(true, forKey: "showShortcutInHUD")
-        WelcomeExperiencePolicy.isScreenshotModeEnabled = true
-        ScreenshotMode.renderStyle = arguments.prefersLiveWindowCapture ? .liveWindow : .synthetic
-        LanguageManager.shared.syncAppleLanguages(for: arguments.language)
-    }
-
-    @MainActor
-    static func seed(store: GroupStore, for arguments: ScreenshotArguments) {
-        let groups = ScreenshotFixtureLibrary.makeGroups()
-        RunningAppQuickAddCatalog.shared.setOverrideApps(ScreenshotFixtureLibrary.makeQuickAddOverrideApps())
-        let export = ScreenshotFixtureLibrary.makeExport(
-            groups: groups,
-            theme: arguments.theme,
-            language: arguments.language
-        )
-        store.applyImport(export)
-
-        if let groupKey = arguments.groupKey,
-           let groupID = ScreenshotFixtureLibrary.groupID(for: groupKey) {
-            store.selectedGroupId = groupID
-        }
-
-        if arguments.scene == .backups {
-            ScreenshotFixtureLibrary.writeBackupFixtures(
-                into: store,
-                groups: groups,
-                language: arguments.language,
-                theme: arguments.theme
-            )
-        }
-
-        switch arguments.scene {
-        case .general:
-            ShortcutCycleURLNavigationState.request(tab: .general)
-        case .group:
-            ShortcutCycleURLNavigationState.request(tab: .groups)
-        case .backups:
-            ShortcutCycleURLNavigationState.request(tab: .general)
-        case .hudHorizontal, .hudGrid, .menuPopover:
-            break
-        }
-    }
-
-    static func shortcutDataMap(for groups: [AppGroup]) -> [String: ShortcutData] {
-        var shortcuts: [String: ShortcutData] = [:]
-
-        for (index, group) in groups.enumerated() where screenshotShortcutKeys.indices.contains(index) {
-            let shortcut = KeyboardShortcuts.Shortcut(screenshotShortcutKeys[index], modifiers: [.option])
-            shortcuts[group.id.uuidString] = ShortcutData(
-                carbonKeyCode: shortcut.carbonKeyCode,
-                carbonModifiers: shortcut.carbonModifiers
-            )
-        }
-
-        return shortcuts
-    }
-
-    @MainActor
-    static func capture(window: NSWindow, arguments: ScreenshotArguments) async throws {
-        prepare(window: window)
-
-        let screenshot: CGImage
-        do {
-            if let liveSnapshot = try await systemSnapshot(window: window) {
-                screenshot = liveSnapshot
-            } else if let cachedSnapshot = snapshot(window: window) {
-                screenshot = cachedSnapshot
-            } else {
-                throw ScreenshotError.captureFailed("Failed to snapshot screenshot window")
-            }
-        } catch {
-            if let cachedSnapshot = snapshot(window: window) {
-                fputs("Falling back to cached screenshot capture: \(error.localizedDescription)\n", stderr)
-                screenshot = cachedSnapshot
-            } else {
-                throw error
-            }
-        }
-
-        let composited = try compositeToOutputCanvas(cgImage: screenshot, theme: arguments.theme)
-        try write(cgImage: composited, to: arguments.outputURL)
-    }
-
-    #if canImport(ScreenCaptureKit)
-    @MainActor
-    private static func systemSnapshot(window: NSWindow) async throws -> CGImage? {
-        guard #available(macOS 14.4, *) else { return nil }
-
-        let content = try await SCShareableContent.currentProcess
-        guard let scWindow = content.windows.first(where: { $0.windowID == CGWindowID(window.windowNumber) }) else {
-            return nil
-        }
-
-        let scale = max(window.backingScaleFactor, 1)
-        let configuration = SCStreamConfiguration()
-        configuration.width = Int(window.frame.width * scale)
-        configuration.height = Int(window.frame.height * scale)
-        configuration.showsCursor = false
-        configuration.scalesToFit = true
-        configuration.captureResolution = .best
-        configuration.ignoreShadowsSingleWindow = false
-        configuration.shouldBeOpaque = false
-
-        let backgroundColor = CGColor(gray: 0, alpha: 0)
-        configuration.backgroundColor = backgroundColor
-
-        let filter = SCContentFilter(desktopIndependentWindow: scWindow)
-
-        return try await withCheckedThrowingContinuation { continuation in
-            SCScreenshotManager.captureImage(contentFilter: filter, configuration: configuration) { image, error in
-                if let image {
-                    continuation.resume(returning: image)
-                    return
-                }
-
-                if let error {
-                    continuation.resume(throwing: error)
-                    return
-                }
-
-                continuation.resume(throwing: ScreenshotError.captureFailed("ScreenCaptureKit returned no image"))
-            }
-        }
-    }
-    #else
-    @MainActor
-    private static func systemSnapshot(window: NSWindow) async throws -> CGImage? {
-        nil
-    }
-    #endif
-
-    @MainActor
-    static func prepareForExternalCapture(window: NSWindow, arguments: ScreenshotArguments) throws {
-        prepare(window: window)
-
-        guard let windowInfoURL = arguments.windowInfoURL else { return }
-        let directory = windowInfoURL.deletingLastPathComponent()
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-        let payload = ScreenshotWindowInfo(windowNumber: window.windowNumber)
-        let data = try JSONEncoder().encode(payload)
-        try data.write(to: windowInfoURL, options: .atomic)
-    }
-
-    @MainActor
-    private static func prepare(window: NSWindow) {
-        window.orderFrontRegardless()
-        if window.canBecomeKey {
-            window.makeKeyAndOrderFront(nil)
-        }
-        if window.canBecomeMain {
-            window.makeMain()
-        }
-        window.displayIfNeeded()
-        window.contentView?.layoutSubtreeIfNeeded()
-        window.contentView?.displayIfNeeded()
-        NSApp.activate(ignoringOtherApps: true)
-        NSRunningApplication.current.activate(options: [.activateAllWindows])
-    }
-
-    static func configure(window: NSWindow, for arguments: ScreenshotArguments) {
-        let size = NSSize(width: windowSize.width, height: windowSize.height)
-        let usesLiveWindowChrome = arguments.prefersLiveWindowCapture && !arguments.scene.hidesWindowChrome
-        if let screenFrame = window.screen?.visibleFrame ?? NSScreen.main?.visibleFrame {
-            let origin = NSPoint(
-                x: screenFrame.midX - (size.width / 2),
-                y: screenFrame.midY - (size.height / 2)
-            )
-            window.setFrame(NSRect(origin: origin, size: size), display: true)
-        } else {
-            window.setFrame(NSRect(origin: .zero, size: size), display: true)
-            window.center()
-        }
-
-        window.appearance = {
-            guard let colorScheme = arguments.theme.colorScheme else { return nil }
-
-            switch colorScheme {
-            case .dark:
-                return NSAppearance(named: .darkAqua)
-            case .light:
-                return NSAppearance(named: .aqua)
-            @unknown default:
-                return nil
-            }
-        }()
-        window.titleVisibility = .hidden
-        window.titlebarAppearsTransparent = arguments.scene.hidesWindowChrome || !usesLiveWindowChrome
-        window.isMovableByWindowBackground = false
-        window.isOpaque = !arguments.scene.hidesWindowChrome
-        window.backgroundColor = arguments.scene.hidesWindowChrome ? .clear : .windowBackgroundColor
-        window.hasShadow = !arguments.scene.hidesWindowChrome
-        window.sharingType = .readOnly
-        if usesLiveWindowChrome {
-            window.toolbarStyle = .unified
-        }
-
-        let buttons: [NSWindow.ButtonType] = [.closeButton, .miniaturizeButton, .zoomButton]
-        for buttonType in buttons {
-            guard let button = window.standardWindowButton(buttonType) else { continue }
-            button.isHidden = !usesLiveWindowChrome
-            button.showsBorderOnlyWhileMouseInside = false
-            button.needsDisplay = true
-        }
-
-        window.standardWindowButton(.closeButton)?.superview?.needsDisplay = true
-
-        if arguments.scene.hidesWindowChrome {
-            window.styleMask.insert(.fullSizeContentView)
-        }
-    }
-
-    static func ensureOutputDirectory(for outputURL: URL) throws {
-        let directory = outputURL.deletingLastPathComponent()
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-    }
-
-    static func write(cgImage: CGImage, to outputURL: URL) throws {
-        try ensureOutputDirectory(for: outputURL)
-        let representation = NSBitmapImageRep(cgImage: cgImage)
-        guard let data = representation.representation(using: .png, properties: [:]) else {
-            throw ScreenshotError.captureFailed("Failed to encode PNG data")
-        }
-        try data.write(to: outputURL, options: .atomic)
-    }
-
-    @MainActor
-    private static func snapshot(window: NSWindow) -> CGImage? {
-        guard let targetView = window.contentView?.superview ?? window.contentView else {
-            return nil
-        }
-
-        targetView.layoutSubtreeIfNeeded()
-        let bounds = targetView.bounds
-        guard let rep = targetView.bitmapImageRepForCachingDisplay(in: bounds) else {
-            return nil
-        }
-        targetView.cacheDisplay(in: bounds, to: rep)
-        return rep.cgImage
-    }
-
-    private static func compositeToOutputCanvas(cgImage: CGImage, theme: AppTheme) throws -> CGImage {
-        let outputWidth = Int(outputPixelSize.width)
-        let outputHeight = Int(outputPixelSize.height)
-
-        guard let context = CGContext(
-            data: nil,
-            width: outputWidth,
-            height: outputHeight,
-            bitsPerComponent: 8,
-            bytesPerRow: 0,
-            space: CGColorSpaceCreateDeviceRGB(),
-            bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
-        ) else {
-            throw ScreenshotError.captureFailed("Failed to create output CGContext")
-        }
-
-        let fillColor = theme == .dark ? NSColor.black : NSColor.white
-        context.setFillColor(fillColor.cgColor)
-        context.fill(CGRect(x: 0, y: 0, width: outputWidth, height: outputHeight))
-
-        let originX = (outputWidth - cgImage.width) / 2
-        let originY = (outputHeight - cgImage.height) / 2
-        context.draw(
-            cgImage,
-            in: CGRect(x: originX, y: originY, width: cgImage.width, height: cgImage.height)
-        )
-
-        guard let outputImage = context.makeImage() else {
-            throw ScreenshotError.captureFailed("Failed to create output CGImage")
-        }
-
-        return outputImage
-    }
-}
-
-private enum ScreenshotError: LocalizedError {
-    case captureFailed(String)
-
-    var errorDescription: String? {
-        switch self {
-        case .captureFailed(let message):
-            return message
-        }
-    }
-}
-
-private enum ScreenshotFixtureLibrary {
-    private struct FixtureApp {
-        let name: String
-        let bundleID: String
-    }
-
-    private struct FixtureGroup {
-        let key: ScreenshotFixtureGroupKey
-        let id: UUID
-        let name: String
-        let apps: [FixtureApp]
-        let isEnabled: Bool
-        let openAppIfNeeded: Bool
-    }
-
-    private static let fixtureGroups: [FixtureGroup] = [
-        FixtureGroup(
-            key: .info,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000001")!,
-            name: "Info",
-            apps: [
-                .init(name: "Weather", bundleID: "com.apple.weather"),
-                .init(name: "Maps", bundleID: "com.apple.Maps"),
-                .init(name: "Stocks", bundleID: "com.apple.stocks"),
-                .init(name: "News", bundleID: "com.apple.news")
-            ],
-            isEnabled: true,
-            openAppIfNeeded: true
-        ),
-        FixtureGroup(
-            key: .communication,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000002")!,
-            name: "Communication",
-            apps: [
-                .init(name: "Messages", bundleID: "com.apple.MobileSMS"),
-                .init(name: "FaceTime", bundleID: "com.apple.FaceTime"),
-                .init(name: "Mail", bundleID: "com.apple.mail"),
-                .init(name: "Contacts", bundleID: "com.apple.AddressBook")
-            ],
-            isEnabled: true,
-            openAppIfNeeded: false
-        ),
-        FixtureGroup(
-            key: .productivity,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000003")!,
-            name: "Productivity",
-            apps: [
-                .init(name: "Calendar", bundleID: "com.apple.iCal"),
-                .init(name: "Reminders", bundleID: "com.apple.reminders"),
-                .init(name: "Notes", bundleID: "com.apple.Notes"),
-                .init(name: "Freeform", bundleID: "com.apple.freeform"),
-                .init(name: "Preview", bundleID: "com.apple.Preview")
-            ],
-            isEnabled: true,
-            openAppIfNeeded: false
-        ),
-        FixtureGroup(
-            key: .utilities,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000004")!,
-            name: "Utilities",
-            apps: [
-                .init(name: "Terminal", bundleID: "com.apple.Terminal"),
-                .init(name: "Activity Monitor", bundleID: "com.apple.ActivityMonitor"),
-                .init(name: "Console", bundleID: "com.apple.Console"),
-                .init(name: "System Settings", bundleID: "com.apple.systempreferences"),
-                .init(name: "Calculator", bundleID: "com.apple.calculator"),
-                .init(name: "Shortcuts", bundleID: "com.apple.shortcuts"),
-                .init(name: "App Store", bundleID: "com.apple.AppStore"),
-                .init(name: "QuickTime Player", bundleID: "com.apple.QuickTimePlayerX")
-            ],
-            isEnabled: true,
-            openAppIfNeeded: true
-        ),
-        FixtureGroup(
-            key: .media,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000005")!,
-            name: "Media",
-            apps: [
-                .init(name: "Music", bundleID: "com.apple.Music"),
-                .init(name: "TV", bundleID: "com.apple.TV"),
-                .init(name: "Podcasts", bundleID: "com.apple.podcasts"),
-                .init(name: "Photos", bundleID: "com.apple.Photos"),
-                .init(name: "Books", bundleID: "com.apple.iBooksX")
-            ],
-            isEnabled: false,
-            openAppIfNeeded: false
-        ),
-        FixtureGroup(
-            key: .manyApps,
-            id: UUID(uuidString: "10000000-0000-0000-0000-000000000006")!,
-            name: "Many Apps",
-            apps: [
-                .init(name: "Weather", bundleID: "com.apple.weather"),
-                .init(name: "Maps", bundleID: "com.apple.Maps"),
-                .init(name: "Music", bundleID: "com.apple.Music"),
-                .init(name: "TV", bundleID: "com.apple.TV"),
-                .init(name: "Photos", bundleID: "com.apple.Photos"),
-                .init(name: "Notes", bundleID: "com.apple.Notes"),
-                .init(name: "Terminal", bundleID: "com.apple.Terminal"),
-                .init(name: "System Settings", bundleID: "com.apple.systempreferences"),
-                .init(name: "App Store", bundleID: "com.apple.AppStore"),
-                .init(name: "Preview", bundleID: "com.apple.Preview")
-            ],
-            isEnabled: true,
-            openAppIfNeeded: true
-        )
-    ]
-
-    static func makeGroups() -> [AppGroup] {
-        fixtureGroups.map { fixtureGroup in
-            AppGroup(
-                id: fixtureGroup.id,
-                name: fixtureGroup.name,
-                apps: fixtureGroup.apps.map(makeAppItem),
-                isEnabled: fixtureGroup.isEnabled,
-                openAppIfNeeded: fixtureGroup.openAppIfNeeded,
-                lastModified: Date(timeIntervalSince1970: 1_735_689_600)
-            )
-        }
-    }
-
-    static func groupID(for key: ScreenshotFixtureGroupKey) -> UUID? {
-        fixtureGroups.first(where: { $0.key == key })?.id
-    }
-
-    static func hudItems(for key: ScreenshotFixtureGroupKey) -> [HUDAppItem] {
-        guard let group = fixtureGroups.first(where: { $0.key == key }) else { return [] }
-        return group.apps.map { app in
-            HUDAppItem(
-                id: app.bundleID,
-                name: app.name,
-                icon: icon(for: app.bundleID),
-                isRunning: true
-            )
-        }
-    }
-
-    static func hudCaptureState(for scene: ScreenshotScene, key: ScreenshotFixtureGroupKey) -> (items: [HUDAppItem], activeItemID: String, shortcut: String?) {
-        let items = hudItems(for: key)
-        let activeItemID: String
-
-        switch scene {
-        case .hudGrid:
-            activeItemID = items.dropFirst(3).first?.id ?? items.first?.id ?? ""
-        default:
-            activeItemID = items.first?.id ?? ""
-        }
-
-        let shortcut: String?
-        if let index = items.firstIndex(where: { $0.id == activeItemID }) {
-            shortcut = "⌥\(index + 1)"
-        } else {
-            shortcut = nil
-        }
-
-        return (items, activeItemID, shortcut)
-    }
-
-    static func menuRunningBundleIDs() -> Set<String> {
-        Set(
-            fixtureGroups
-                .filter(\.isEnabled)
-                .flatMap { group in group.apps.map(\.bundleID) }
-        )
-    }
-
-    static func makeQuickAddOverrideApps() -> [AppItem] {
-        var seenBundleIDs = Set<String>()
-        return fixtureGroups
-            .flatMap(\.apps)
-            .filter { seenBundleIDs.insert($0.bundleID).inserted }
-            .map(makeAppItem)
-            .sorted {
-                let lhs = $0.name.folding(options: [.caseInsensitive, .diacriticInsensitive], locale: .current)
-                let rhs = $1.name.folding(options: [.caseInsensitive, .diacriticInsensitive], locale: .current)
-                if lhs == rhs {
-                    return $0.bundleIdentifier < $1.bundleIdentifier
-                }
-                return lhs < rhs
-            }
-    }
-
-    static func makeExport(groups: [AppGroup], theme: AppTheme, language: String) -> SettingsExport {
-        SettingsExport(
-            groups: groups,
-            settings: AppSettings(
-                showHUD: true,
-                showShortcutInHUD: true,
-                selectedLanguage: language,
-                appTheme: theme.rawValue
-            ),
-            shortcuts: ScreenshotRuntime.shortcutDataMap(for: groups)
-        )
-    }
-
-    @MainActor
-    static func writeBackupFixtures(
-        into store: GroupStore,
-        groups: [AppGroup],
-        language: String,
-        theme: AppTheme
-    ) {
-        let fileManager = FileManager.default
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-
-        try? fileManager.removeItem(at: store.backupDirectory)
-        try? fileManager.createDirectory(at: store.backupDirectory, withIntermediateDirectories: true)
-
-        let comparisonGroups = groups.enumerated().map { index, group -> AppGroup in
-            guard index == 0 else { return group }
-            var updatedGroup = group
-            updatedGroup.apps = Array(group.apps.dropLast())
-            return updatedGroup
-        }
-
-        let exports: [(name: String, export: SettingsExport, createdAt: Date)] = [
-            (
-                "backup 2026-02-02 17-46-00.json",
-                makeExport(groups: comparisonGroups, theme: theme == .dark ? .light : .dark, language: language),
-                Date(timeIntervalSince1970: 1_770_074_760)
-            ),
-            (
-                "backup 2026-02-02 17-52-00.json",
-                makeExport(groups: groups, theme: theme, language: language),
-                Date(timeIntervalSince1970: 1_770_075_120)
-            )
-        ]
-
-        for exportFile in exports {
-            let fileURL = store.backupDirectory.appendingPathComponent(exportFile.name)
-            guard let data = try? encoder.encode(exportFile.export) else { continue }
-            try? data.write(to: fileURL, options: .atomic)
-            try? fileManager.setAttributes([.creationDate: exportFile.createdAt], ofItemAtPath: fileURL.path)
-        }
-    }
-
-    private static func makeAppItem(from definition: FixtureApp) -> AppItem {
-        let appURL = NSWorkspace.shared.urlForApplication(withBundleIdentifier: definition.bundleID)
-        return AppItem(
-            bundleIdentifier: definition.bundleID,
-            name: definition.name,
-            iconPath: appURL?.path
-        )
-    }
-
-    private static func icon(for bundleID: String) -> NSImage {
-        if let appURL = NSWorkspace.shared.urlForApplication(withBundleIdentifier: bundleID) {
-            return NSWorkspace.shared.icon(forFile: appURL.path)
-        }
-
-        return NSWorkspace.shared.icon(for: .applicationBundle)
-    }
-}
-
-private struct ScreenshotSceneContainerView: View {
-    @EnvironmentObject private var store: GroupStore
-
-    let arguments: ScreenshotArguments
-    let localeObserver: LocaleObserver
-
-    var body: some View {
-        Group {
-            if ScreenshotMode.usesSyntheticControls {
-                content
-                    .tint(.accentColor)
-            } else {
-                content
-            }
-        }
-            .environment(\.controlActiveState, .key)
-            .overlay(alignment: .topLeading) {
-                if arguments.scene.showsTrafficLightsOverlay && ScreenshotMode.usesSyntheticChrome {
-                    ScreenshotTrafficLightsOverlay()
-                        .padding(.top, 14)
-                        .padding(.leading, 16)
-                }
-            }
-    }
-
-    @ViewBuilder
-    private var content: some View {
-        switch arguments.scene {
-        case .general:
-            if arguments.prefersLiveWindowCapture {
-                MainView()
-                    .environmentObject(store)
-                    .environmentObject(localeObserver)
-            } else {
-                ScreenshotGeneralSceneView(localeObserver: localeObserver)
-                    .environmentObject(store)
-            }
-        case .group:
-            if arguments.prefersLiveWindowCapture {
-                MainView()
-                    .environmentObject(store)
-                    .environmentObject(localeObserver)
-            } else {
-                ScreenshotGroupSceneView()
-                    .environmentObject(store)
-            }
-        case .backups:
-            if arguments.prefersLiveWindowCapture {
-                MainView()
-                    .environmentObject(store)
-                    .environmentObject(localeObserver)
-            } else {
-                ScreenshotBackupSceneView(localeObserver: localeObserver)
-                    .environmentObject(store)
-            }
-        case .hudHorizontal, .hudGrid:
-            ScreenshotHUDSceneView(arguments: arguments)
-        case .menuPopover:
-            ScreenshotMenuPopoverSceneView(arguments: arguments)
-                .environmentObject(store)
-        }
-    }
-}
-
-@MainActor
-private enum ScreenshotWindowLifecycle {
-    static var store: GroupStore?
-    static var controller: NSWindowController?
-
-    static func configure(store: GroupStore) {
-        self.store = store
-    }
-
-    static func present(arguments: ScreenshotArguments) {
-        guard let store else { return }
-
-        if arguments.prefersLiveWindowCapture {
-            switch arguments.scene {
-            case .general:
-                ShortcutCycleURLNavigationState.request(tab: .general)
-            case .group:
-                ShortcutCycleURLNavigationState.request(tab: .groups)
-            case .backups:
-                ShortcutCycleURLNavigationState.requestBackupBrowser()
-            case .hudHorizontal, .hudGrid, .menuPopover:
-                break
-            }
-        }
-
-        let captureWindow: NSWindow
-        switch arguments.scene {
-        case .hudHorizontal where arguments.prefersLiveWindowCapture,
-             .hudGrid where arguments.prefersLiveWindowCapture:
-            self.controller = nil
-            captureWindow = presentLiveHUDWindow(arguments: arguments)
-        case .menuPopover where arguments.prefersLiveWindowCapture:
-            captureWindow = presentLiveMenuPopoverWindow(arguments: arguments, store: store)
-        default:
-            captureWindow = presentSceneWindow(arguments: arguments, store: store)
-        }
-
-        let timeout = arguments.windowInfoURL == nil
-            ? max(arguments.captureDelay + 4, 8.0)
-            : max(arguments.captureDelay + 20, 25.0)
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + arguments.captureDelay) {
-            Task { @MainActor in
-                do {
-                    if arguments.windowInfoURL != nil {
-                        try ScreenshotRuntime.prepareForExternalCapture(window: captureWindow, arguments: arguments)
-                    } else {
-                        try await ScreenshotRuntime.capture(window: captureWindow, arguments: arguments)
-                        exit(0)
-                    }
-                } catch {
-                    fputs("Screenshot capture failed: \(error.localizedDescription)\n", stderr)
-                    exit(1)
-                }
-            }
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + timeout) {
-            fputs("Screenshot capture timed out for scene \(arguments.scene.rawValue)\n", stderr)
-            exit(2)
-        }
-    }
-
-    private static func presentSceneWindow(arguments: ScreenshotArguments, store: GroupStore) -> NSWindow {
-        let rootView = ScreenshotSceneContainerView(
-            arguments: arguments,
-            localeObserver: LocaleObserver()
-        )
-        .environmentObject(store)
-
-        let hostingController = NSHostingController(rootView: rootView)
-        let styleMask: NSWindow.StyleMask = arguments.scene.hidesWindowChrome
-            ? [.borderless]
-            : [.titled, .closable, .miniaturizable]
-        let window = NSWindow(
-            contentRect: NSRect(
-                x: 0,
-                y: 0,
-                width: ScreenshotRuntime.windowSize.width,
-                height: ScreenshotRuntime.windowSize.height
-            ),
-            styleMask: styleMask,
-            backing: .buffered,
-            defer: false
-        )
-        window.contentViewController = hostingController
-        window.isReleasedWhenClosed = false
-        window.title = arguments.prefersLiveWindowCapture ? "Shortcut Cycle" : "ShortcutCycle Screenshot"
-        window.isMovable = false
-
-        let controller = NSWindowController(window: window)
-        self.controller = controller
-        controller.showWindow(nil)
-        NSApp.activate(ignoringOtherApps: true)
-        ScreenshotRuntime.configure(window: window, for: arguments)
-        return window
-    }
-
-    private static func presentLiveHUDWindow(arguments: ScreenshotArguments) -> NSWindow {
-        let groupKey = arguments.groupKey ?? arguments.scene.defaultGroupKey ?? .info
-        let hudCapture = ScreenshotFixtureLibrary.hudCaptureState(for: arguments.scene, key: groupKey)
-
-        guard let window = HUDManager.shared.presentScreenshotHUD(
-            items: hudCapture.items,
-            activeAppId: hudCapture.activeItemID,
-            shortcut: hudCapture.shortcut
-        ) else {
-            fatalError("Failed to create HUD screenshot window")
-        }
-
-        return window
-    }
-
-    private static func presentLiveMenuPopoverWindow(arguments: ScreenshotArguments, store: GroupStore) -> NSWindow {
-        let selectedGroupID: UUID? = {
-            switch arguments.menuVariant {
-            case .default:
-                return nil
-            case .selected:
-                return arguments.groupKey.flatMap(ScreenshotFixtureLibrary.groupID(for:))
-                    ?? ScreenshotFixtureLibrary.groupID(for: .utilities)
-            }
-        }()
-
-        let rootView = MenuBarView(
-            selectedLanguage: arguments.language,
-            screenshotHighlightedGroupID: selectedGroupID,
-            screenshotRunningBundleIDs: ScreenshotFixtureLibrary.menuRunningBundleIDs(),
-            screenshotLaunchAtLogin: true,
-            screenshotThemeOverride: arguments.theme
-        )
-        .environment(\.controlActiveState, .key)
-        .environmentObject(store)
-
-        let hostingController = NSHostingController(rootView: rootView)
-        let window = ScreenshotPopoverCaptureWindow(
-            contentRect: NSRect(origin: .zero, size: NSSize(width: 280, height: 400)),
-            styleMask: [.borderless],
-            backing: .buffered,
-            defer: false
-        )
-        window.contentViewController = hostingController
-        window.isReleasedWhenClosed = false
-        window.isMovable = false
-        window.backgroundColor = .clear
-        window.isOpaque = false
-        window.hasShadow = false
-        window.ignoresMouseEvents = true
-        window.sharingType = .readOnly
-
-        hostingController.view.layoutSubtreeIfNeeded()
-        let fittingSize = hostingController.view.fittingSize
-        if fittingSize.width > 0, fittingSize.height > 0 {
-            window.setContentSize(fittingSize)
-        }
-
-        let controller = NSWindowController(window: window)
-        self.controller = controller
-        controller.showWindow(nil)
-        window.center()
-        NSApp.activate(ignoringOtherApps: true)
-        window.orderFrontRegardless()
-        window.displayIfNeeded()
-        window.contentView?.layoutSubtreeIfNeeded()
-        window.contentView?.displayIfNeeded()
-        return window
-    }
-}
-
-private final class ScreenshotPopoverCaptureWindow: NSWindow {
-    override var canBecomeKey: Bool { true }
-    override var canBecomeMain: Bool { true }
-}
-
-private struct ScreenshotBackupSceneView: View {
-    @EnvironmentObject private var store: GroupStore
-    @Environment(\.colorScheme) private var colorScheme
-
-    let localeObserver: LocaleObserver
-
-    var body: some View {
-        ZStack {
-            ScreenshotGeneralSceneView(localeObserver: localeObserver)
-                .environmentObject(store)
-                .allowsHitTesting(false)
-
-            Color.black.opacity(colorScheme == .dark ? 0.10 : 0.08)
-                .ignoresSafeArea()
-
-            BackupBrowserView()
-                .environmentObject(store)
-                .frame(width: 650, height: 450)
-                .background(
-                    RoundedRectangle(cornerRadius: 22, style: .continuous)
-                        .fill(Color(nsColor: .windowBackgroundColor))
-                )
-                .clipShape(RoundedRectangle(cornerRadius: 22, style: .continuous))
-                .shadow(
-                    color: .black.opacity(colorScheme == .dark ? 0.38 : 0.18),
-                    radius: colorScheme == .dark ? 42 : 30,
-                    x: 0,
-                    y: colorScheme == .dark ? 18 : 12
-                )
-        }
-    }
-}
-
-private struct ScreenshotGeneralSceneView: View {
-    @EnvironmentObject private var store: GroupStore
-    @Environment(\.colorScheme) private var colorScheme
-    @AppStorage("selectedLanguage") private var selectedLanguage = "system"
-
-    let localeObserver: LocaleObserver
-
-    var body: some View {
-        ScreenshotSettingsWindowChrome(selectedLanguage: selectedLanguage, selectedTab: .general) {
-            GeneralSettingsView()
-                .environmentObject(store)
-                .environmentObject(localeObserver)
-        }
-        .background(SettingsChromePalette.windowBackground(for: colorScheme))
-    }
-}
-
-private struct ScreenshotTrafficLightsOverlay: View {
-    private let colors: [Color] = [
-        Color(red: 1.0, green: 0.37, blue: 0.34),
-        Color(red: 1.0, green: 0.74, blue: 0.18),
-        Color(red: 0.18, green: 0.82, blue: 0.35)
-    ]
-
-    var body: some View {
-        HStack(spacing: 8) {
-            ForEach(colors.indices, id: \.self) { index in
-                Circle()
-                    .fill(colors[index])
-                    .frame(width: 12, height: 12)
-                    .overlay(
-                        Circle()
-                            .stroke(Color.black.opacity(0.12), lineWidth: 0.5)
-                    )
-            }
-        }
-    }
-}
-
-private struct ScreenshotGroupSceneView: View {
-    @EnvironmentObject private var store: GroupStore
-    @AppStorage("selectedLanguage") private var selectedLanguage = "system"
-    @Environment(\.colorScheme) private var colorScheme
-
-    private var selectedGroupID: UUID? {
-        store.selectedGroupId ?? store.groups.first?.id
-    }
-
-    var body: some View {
-        ScreenshotSettingsWindowChrome(selectedLanguage: selectedLanguage, selectedTab: .groups) {
-            HStack(spacing: 0) {
-                ScreenshotGroupSidebar(selectedGroupID: selectedGroupID)
-                    .frame(width: 220)
-
-                Divider()
-
-                if let selectedGroupID {
-                    GroupEditView(groupId: selectedGroupID)
-                        .environmentObject(store)
-                } else {
-                    Color.clear
-                }
-            }
-        }
-        .background(SettingsChromePalette.windowBackground(for: colorScheme))
-    }
-}
-
-private struct ScreenshotSettingsWindowChrome<Content: View>: View {
-    let selectedLanguage: String
-    let selectedTab: ScreenshotSettingsTabsView.Tab
-    @ViewBuilder let content: () -> Content
-
-    var body: some View {
-        VStack(spacing: 0) {
-            HStack {
-                Spacer()
-                ScreenshotSettingsTabsView(selectedLanguage: selectedLanguage, selectedTab: selectedTab)
-                Spacer()
-            }
-            .frame(height: 42)
-            .background(Color(nsColor: .windowBackgroundColor))
-
-            Divider()
-
-            content()
-        }
-    }
-}
-
-private struct ScreenshotSettingsTabsView: View {
-    enum Tab {
-        case groups
-        case general
-    }
-
-    let selectedLanguage: String
-    let selectedTab: Tab
-
-    var body: some View {
-        HStack(spacing: 4) {
-            tab("Groups".localized(language: selectedLanguage), isSelected: selectedTab == .groups)
-            tab("General".localized(language: selectedLanguage), isSelected: selectedTab == .general)
-        }
-        .padding(4)
-        .background(
-            Capsule(style: .continuous)
-                .fill(Color.secondary.opacity(0.12))
-        )
-    }
-
-    private func tab(_ title: String, isSelected: Bool) -> some View {
-        Text(title)
-            .font(.subheadline.weight(.semibold))
-            .foregroundStyle(isSelected ? .primary : .secondary)
-            .padding(.horizontal, 18)
-            .padding(.vertical, 6)
-            .background(
-                Capsule(style: .continuous)
-                    .fill(isSelected ? Color(nsColor: .controlBackgroundColor) : .clear)
-            )
-    }
-}
-
-private struct ScreenshotGroupSidebar: View {
-    @EnvironmentObject private var store: GroupStore
-    @Environment(\.colorScheme) private var colorScheme
-
-    let selectedGroupID: UUID?
-
-    var body: some View {
-        VStack(spacing: 0) {
-            ScrollView {
-                LazyVStack(spacing: 6) {
-                    ForEach(store.groups) { group in
-                        ScreenshotGroupSidebarRow(
-                            group: group,
-                            isSelected: group.id == selectedGroupID
-                        )
-                    }
-                }
-                .padding(10)
-            }
-
-            Divider()
-
-            HStack {
-                Label("Add Group", systemImage: "plus")
-                    .font(.body)
-                Spacer()
-            }
-            .padding(10)
-        }
-        .background(SettingsChromePalette.sidebarBackground(for: colorScheme))
-    }
-}
-
-private struct ScreenshotGroupSidebarRow: View {
-    @Environment(\.colorScheme) private var colorScheme
-    @AppStorage("selectedLanguage") private var selectedLanguage = "system"
-
-    let group: AppGroup
-    let isSelected: Bool
-
-    var body: some View {
-        HStack(spacing: 8) {
-            if ScreenshotMode.usesSyntheticControls {
-                ScreenshotAccentSwitch(isOn: group.isEnabled, size: .mini)
-            } else {
-                Toggle("", isOn: .constant(group.isEnabled))
-                    .labelsHidden()
-                    .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
-                    .controlSize(.mini)
-            }
-
-            Image(systemName: "folder.fill")
-                .foregroundStyle(group.isEnabled ? AnyShapeStyle(Color.accentColor) : AnyShapeStyle(.secondary))
-
-            VStack(alignment: .leading, spacing: 2) {
-                Text(group.name)
-                    .font(.subheadline.weight(.medium))
-                    .foregroundStyle(group.isEnabled ? .primary : .secondary)
-                    .lineLimit(1)
-
-                if let shortcut = group.shortcutDisplayString {
-                    Text(shortcut)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                        .padding(.horizontal, 6)
-                        .padding(.vertical, 2)
-                        .background(SettingsChromePalette.chipFill(for: colorScheme))
-                        .clipShape(Capsule(style: .continuous))
-                } else {
-                    Text("No shortcut".localized(language: selectedLanguage))
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                        .lineLimit(1)
-                }
-            }
-
-            Spacer(minLength: 0)
-
-            Text("\(group.apps.count)")
-                .font(colorScheme == .dark ? .caption.weight(.semibold) : .caption)
-                .foregroundStyle(colorScheme == .dark ? AnyShapeStyle(.secondary) : AnyShapeStyle(.white))
-                .padding(.horizontal, 6)
-                .padding(.vertical, 2)
-                .background(Capsule().fill(SettingsChromePalette.badgeFill(for: colorScheme)))
-        }
-        .padding(.horizontal, 8)
-        .padding(.vertical, 8)
-        .background(
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .fill(
-                    isSelected
-                        ? SettingsChromePalette.neutralHoverFill(for: colorScheme)
-                        : Color.clear
-                )
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .stroke(
-                    isSelected ? SettingsChromePalette.neutralHoverBorder(for: colorScheme) : Color.clear,
-                    lineWidth: 1
-                )
-        )
-        .opacity(group.isEnabled ? 1.0 : 0.68)
-    }
-}
-
-private struct ScreenshotHUDSceneView: View {
-    let arguments: ScreenshotArguments
-
-    private var items: [HUDAppItem] {
-        ScreenshotFixtureLibrary.hudItems(for: arguments.groupKey ?? .info)
-    }
-
-    private var activeItemID: String {
-        switch arguments.scene {
-        case .hudGrid:
-            return items.dropFirst(3).first?.id ?? items.first?.id ?? ""
-        default:
-            return items.first?.id ?? ""
-        }
-    }
-
-    var body: some View {
-        ZStack {
-            ScreenshotBackdrop(theme: arguments.theme, backgroundURL: arguments.backgroundURL)
-            AppSwitcherHUDView(
-                apps: items,
-                activeAppId: activeItemID,
-                shortcutString: "⌥\(items.firstIndex(where: { $0.id == activeItemID }).map { String($0 + 1) } ?? "1")"
-            )
-            .fixedSize()
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .ignoresSafeArea()
-        .preferredColorScheme(arguments.theme.colorScheme)
-    }
-}
-
-private struct ScreenshotBackdrop: View {
-    let theme: AppTheme
-    let backgroundURL: URL?
-
-    private var backgroundImage: NSImage? {
-        guard let backgroundURL else { return nil }
-        return NSImage(contentsOf: backgroundURL)
-    }
-
-    var body: some View {
-        GeometryReader { proxy in
-            ZStack {
-                if let backgroundImage {
-                    Image(nsImage: backgroundImage)
-                        .resizable()
-                        .scaledToFill()
-                        .frame(width: proxy.size.width, height: proxy.size.height)
-                        .clipped()
-                } else {
-                    LinearGradient(
-                        colors: gradientColors,
-                        startPoint: .topLeading,
-                        endPoint: .bottomTrailing
-                    )
-                    ForEach(backgroundShapes.indices, id: \.self) { index in
-                        Circle()
-                            .fill(backgroundShapes[index].color)
-                            .frame(width: backgroundShapes[index].size, height: backgroundShapes[index].size)
-                            .blur(radius: 40)
-                            .offset(backgroundShapes[index].offset)
-                    }
-                }
-            }
-        }
-    }
-
-    private var gradientColors: [Color] {
-        if theme == .dark {
-            return [
-                Color(red: 0.08, green: 0.10, blue: 0.16),
-                Color(red: 0.12, green: 0.16, blue: 0.25),
-                Color(red: 0.10, green: 0.23, blue: 0.29)
-            ]
-        }
-
-        return [
-            Color(red: 0.91, green: 0.96, blue: 1.0),
-            Color(red: 0.84, green: 0.92, blue: 0.98),
-            Color(red: 0.93, green: 0.96, blue: 0.89)
-        ]
-    }
-
-    private var backgroundShapes: [(size: CGFloat, offset: CGSize, color: Color)] {
-        if theme == .dark {
-            return [
-                (420, CGSize(width: -380, height: -240), Color.cyan.opacity(0.18)),
-                (320, CGSize(width: 360, height: -180), Color.blue.opacity(0.18)),
-                (500, CGSize(width: 280, height: 260), Color.green.opacity(0.12))
-            ]
-        }
-
-        return [
-            (420, CGSize(width: -360, height: -260), Color.white.opacity(0.55)),
-            (300, CGSize(width: 320, height: -160), Color.cyan.opacity(0.22)),
-            (440, CGSize(width: 260, height: 260), Color.green.opacity(0.18))
-        ]
-    }
-}
-
-private struct ScreenshotMenuPopoverSceneView: View {
-    @EnvironmentObject private var store: GroupStore
-
-    let arguments: ScreenshotArguments
-
-    private var selectedGroupID: UUID? {
-        switch arguments.menuVariant {
-        case .default:
-            return nil
-        case .selected:
-            return arguments.groupKey.flatMap(ScreenshotFixtureLibrary.groupID(for:))
-                ?? ScreenshotFixtureLibrary.groupID(for: .utilities)
-        }
-    }
-
-    var body: some View {
-        ZStack {
-            Color(nsColor: arguments.theme == .dark ? .windowBackgroundColor : .underPageBackgroundColor)
-                .ignoresSafeArea()
-
-            VStack {
-                Spacer(minLength: 90)
-                MenuBarView(
-                    selectedLanguage: arguments.language,
-                    screenshotHighlightedGroupID: selectedGroupID,
-                    screenshotLaunchAtLogin: true,
-                    screenshotThemeOverride: arguments.theme
-                )
-                .fixedSize()
-                .environmentObject(store)
-                Spacer()
-            }
-        }
-        .preferredColorScheme(arguments.theme.colorScheme)
-    }
-}
-
-
 @main
 struct ShortcutCycleApp: App {
     @StateObject private var store: GroupStore
     @AppStorage("selectedLanguage") private var selectedLanguage = "system"
-    @AppStorage("appTheme") private var appTheme: AppTheme = .system
     @StateObject private var localeObserver = LocaleObserver()
 
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
 
     init() {
+        #if DEBUG
         if let screenshotArguments = ScreenshotArguments.current {
             ScreenshotRuntime.primeDefaults(for: screenshotArguments)
             LaunchAtLoginManager.setScreenshotOverride(isEnabled: true)
 
-            let screenshotStore = GroupStore(
-                userDefaults: .standard,
-                backupDebounceInterval: 3600,
-                saveDebounceInterval: 0,
-                autoBackupEnabled: false
-            )
+            let screenshotStore = ScreenshotRuntime.makeStore()
             ScreenshotRuntime.seed(store: screenshotStore, for: screenshotArguments)
             ScreenshotWindowLifecycle.configure(store: screenshotStore)
             _store = StateObject(wrappedValue: screenshotStore)
@@ -2151,6 +805,14 @@ struct ShortcutCycleApp: App {
                 ShortcutManager.shared.registerAllShortcuts()
             }
         }
+        #else
+        _store = StateObject(wrappedValue: GroupStore.shared)
+
+        Task { @MainActor in
+            ShortcutManager.shared.registerAllShortcuts()
+        }
+        #endif
+
         // Sync AppleLanguages so third-party bundles (e.g. KeyboardShortcuts) use the
         // correct locale. Must run before any view creates a KeyboardShortcuts.Recorder.
         let selected = UserDefaults.standard.string(forKey: "selectedLanguage") ?? "system"
@@ -2158,11 +820,17 @@ struct ShortcutCycleApp: App {
     }
 
     var body: some Scene {
+        #if DEBUG
+        let isScreenshotMode = ScreenshotArguments.current != nil
+        #else
+        let isScreenshotMode = false
+        #endif
+
         // Menu bar extra
         MenuBarExtra(
             "Shortcut Cycle",
             systemImage: "command.square.fill",
-            isInserted: .constant(ScreenshotArguments.current == nil)
+            isInserted: .constant(!isScreenshotMode)
         ) {
             MenuBarView(selectedLanguage: selectedLanguage)
                 .environmentObject(store)
@@ -2193,12 +861,16 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     private var suppressAutomaticWelcomeForCurrentLaunch = false
 
     func applicationWillFinishLaunching(_ notification: Notification) {
+        #if DEBUG
         if ScreenshotArguments.current != nil {
             NSApp.setActivationPolicy(.regular)
         } else {
             // Run as a menu bar app (no dock icon)
             NSApp.setActivationPolicy(.accessory)
         }
+        #else
+        NSApp.setActivationPolicy(.accessory)
+        #endif
 
         // Register for URL events directly via Apple Events.
         // This is more reliable than application(_:open:) or .onOpenURL,
@@ -2212,7 +884,9 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     }
 
     func applicationDidBecomeActive(_ notification: Notification) {
+        #if DEBUG
         guard ScreenshotArguments.current == nil else { return }
+        #endif
         guard !hasEvaluatedInitialManualActivation else { return }
         hasEvaluatedInitialManualActivation = true
 
@@ -2224,15 +898,19 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     }
 
     func applicationDidFinishLaunching(_ notification: Notification) {
+        #if DEBUG
         guard let screenshotArguments = ScreenshotArguments.current else { return }
 
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.05) {
             ScreenshotWindowLifecycle.present(arguments: screenshotArguments)
         }
+        #endif
     }
 
     func applicationShouldHandleReopen(_ sender: NSApplication, hasVisibleWindows flag: Bool) -> Bool {
+        #if DEBUG
         guard ScreenshotArguments.current == nil else { return false }
+        #endif
         guard !flag else { return false }
         ShortcutCycleURLRouter.openSettingsFromOutsideView()
         return true
@@ -2246,7 +924,9 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         _ event: NSAppleEventDescriptor,
         withReply reply: NSAppleEventDescriptor
     ) {
+        #if DEBUG
         guard ScreenshotArguments.current == nil else { return }
+        #endif
         // Apple Events are delivered before applicationDidBecomeActive on a URL-scheme
         // cold launch, so setting this flag here reliably prevents auto-open of Settings
         // when the launch was triggered externally (e.g. shortcutcycle:// URL).
