+++ ShortcutCycle/ShortcutCycle/Views/Settings/GeneralSettingsView.swift
@@ -37,9 +37,11 @@ struct GeneralSettingsView: View {
     @State private var clipboardImportSummary = ""
     @State private var pendingClipboardExport: SettingsExport?
 
+    #if DEBUG
     private var isScreenshotMode: Bool {
         ScreenshotMode.usesSyntheticControls
     }
+    #endif
 
     var body: some View {
         content
@@ -151,6 +153,7 @@ struct GeneralSettingsView: View {
                 .listRowInsets(EdgeInsets())
                 .padding()
 
+                #if DEBUG
                 if isScreenshotMode {
                     screenshotToggleRow("Show HUD when switching".localized(language: selectedLanguage), isOn: showHUD)
                 } else {
@@ -158,8 +161,14 @@ struct GeneralSettingsView: View {
                         .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
                         .tint(.accentColor)
                 }
+                #else
+                Toggle("Show HUD when switching".localized(language: selectedLanguage), isOn: $showHUD)
+                    .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
+                    .tint(.accentColor)
+                #endif
 
                 if showHUD {
+                    #if DEBUG
                     if isScreenshotMode {
                         screenshotToggleRow("Show shortcut in HUD".localized(language: selectedLanguage), isOn: showShortcutInHUD)
                             .padding(.leading)
@@ -169,6 +178,12 @@ struct GeneralSettingsView: View {
                             .tint(.accentColor)
                             .padding(.leading)
                     }
+                    #else
+                    Toggle("Show shortcut in HUD".localized(language: selectedLanguage), isOn: $showShortcutInHUD)
+                        .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
+                        .tint(.accentColor)
+                        .padding(.leading)
+                    #endif
 
                     Text("Displays the keyboard shortcut used to trigger the switch.".localized(language: selectedLanguage))
                         .font(.caption)
@@ -182,6 +197,7 @@ struct GeneralSettingsView: View {
             }
 
             Section {
+                #if DEBUG
                 if isScreenshotMode {
                     screenshotToggleRow("Open at Login".localized(language: selectedLanguage), isOn: launchAtLogin.isEnabled)
                 } else {
@@ -189,6 +205,11 @@ struct GeneralSettingsView: View {
                         .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
                         .tint(.accentColor)
                 }
+                #else
+                Toggle("Open at Login".localized(language: selectedLanguage), isOn: $launchAtLogin.isEnabled)
+                    .toggleStyle(SwitchToggleStyle(tint: Color(nsColor: .controlAccentColor)))
+                    .tint(.accentColor)
+                #endif
 
                 Picker("Appearance".localized(language: selectedLanguage), selection: $appTheme) {
                     ForEach(AppTheme.allCases) { theme in
@@ -305,6 +326,7 @@ struct GeneralSettingsView: View {
         .background(SettingsChromePalette.windowBackground(for: colorScheme))
     }
 
+    #if DEBUG
     private func screenshotToggleRow(_ title: String, isOn: Bool) -> some View {
         HStack(spacing: 12) {
             Text(title)
@@ -312,6 +334,7 @@ struct GeneralSettingsView: View {
             ScreenshotAccentSwitch(isOn: isOn)
         }
     }
+    #endif
 
     // MARK: - Export/Import Actions
